RavenDB-11275 - Instead of using stack space for buffers, use arrays.
