+++ src/Corax/Queries/QueryContext.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Buffers;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Corax.Queries
+{
+    public static class QueryContext
+    {
+        public static readonly ArrayPool<byte> MatchesPool = ArrayPool<byte>.Create();
+    }
+}
